Added a ReplaySubject<IVersionedEvent> to make EventSourced more amenable to Rx-ification.
This will enable the persistence operations to be moved from repository to object while still maintaining a decoupling from that store. Proceeding in increments, so this is a first step. See EventSourcedTestFixture for tests of Rx functionality and publication.
Added some initial stubs for auto-discovery and wire-up of messaging publication and handling. Shaving the yak a bit, but having observable event streams will help a lot with event and command publication and invocation.